# definitions/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Swagger Test API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "ben-hx@web.de"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8080",
  "basePath": "/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/user/{user_id}/activate": {
      "post": {
        "tags": [
          "user",
          "activation"
        ],
        "summary": "Activate a unregistered user (only for user-role = 'admin')",
        "description": "",
        "operationId": "activateUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID of the user that needs to be activated",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Activated User response",
            "schema": {
              "$ref": "./definitions/user/UserResponse.json"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "./definitions/error/ErrorResponse.json"
            }
          }
        },
        "x-swagger-router-controller": "user-activation-controller"
      }
    },
    "/user/inactive": {
      "get": {
        "tags": [
          "user",
          "activation"
        ],
        "summary": "Get all inaktive users (only for user-role = 'admin')",
        "description": "",
        "operationId": "activateUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page to filter by",
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "User-count-limit per page to filter by",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Inactive Users response",
            "schema": {
              "$ref": "./definitions/user/UsersResponse.json"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "./definitions/error/ErrorResponse.json"
            }
          }
        },
        "x-swagger-router-controller": "user-activation-controller"
      }
    },
...
    "/changepassword": {
      "post": {
        "tags": [
          "change password"
        ],
        "description": "User can change his password",
        "operationId": "changePassword",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Old and new Password",
            "required": true,
            "schema": {
              "$ref": "./definitions/user/ChangePassword.json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User response",
            "schema": {
              "$ref": "./definitions/user/UserResponse.json"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "./definitions/error/ErrorResponse.json"
            }
          }
        },
        "x-swagger-router-controller": "user-controller"
      }
    },
...
  },
  "securityDefinitions": {
    "user_auth": {
      "type": "basic",
      "description": "HTTP Basic Authentication."
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
...
}